Function t3rn_primitives::bridges::chain_circuit::derive_account_from_gateway_id

pub fn derive_account_from_gateway_id(id: SourceAccount<AccountId>) -> AccountId

We use this to get the account on Circuit (target) which is derived from Gateway (source) account. We do this so we can fund the derived account on Circuit at Genesis to it can pay transaction fees.

The reason we can use the same AccountId type for both chains is because they share the same development seed phrase.

Note that this should only be used for testing.