Trait t3rn_primitives::bridges::chain_circuit::FromCircuitInboundLaneApi

pub trait FromCircuitInboundLaneApi<Block: BlockT>: Core<Block> {
    fn latest_received_nonce(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        lane: LaneId
    ) -> Result<MessageNonce, ApiError> { ... }
    fn latest_received_nonce_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        lane: LaneId
    ) -> Result<MessageNonce, ApiError> { ... }
    fn latest_confirmed_nonce(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        lane: LaneId
    ) -> Result<MessageNonce, ApiError> { ... }
    fn latest_confirmed_nonce_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        lane: LaneId
    ) -> Result<MessageNonce, ApiError> { ... }
    fn unrewarded_relayers_state(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        lane: LaneId
    ) -> Result<UnrewardedRelayersState, ApiError> { ... }
    fn unrewarded_relayers_state_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        lane: LaneId
    ) -> Result<UnrewardedRelayersState, ApiError> { ... }
}

Inbound message lane API for messages sent by Circuit chain.

This API is implemented by runtimes that are receiving messages from Circuit chain, not the Circuit runtime itself.

Provided Methods

fn latest_received_nonce(
&self,
__runtime_api_at_param__: &BlockId<Block>,
lane: LaneId
) -> Result<MessageNonce, ApiError>

Returns nonce of the latest message, received by given lane.

fn latest_received_nonce_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
lane: LaneId
) -> Result<MessageNonce, ApiError>

Returns nonce of the latest message, received by given lane.

fn latest_confirmed_nonce(
&self,
__runtime_api_at_param__: &BlockId<Block>,
lane: LaneId
) -> Result<MessageNonce, ApiError>

Nonce of latest message that has been confirmed to the bridged chain.

fn latest_confirmed_nonce_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
lane: LaneId
) -> Result<MessageNonce, ApiError>

Nonce of latest message that has been confirmed to the bridged chain.

fn unrewarded_relayers_state(
&self,
__runtime_api_at_param__: &BlockId<Block>,
lane: LaneId
) -> Result<UnrewardedRelayersState, ApiError>

State of the unrewarded relayers set at given lane.

fn unrewarded_relayers_state_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
lane: LaneId
) -> Result<UnrewardedRelayersState, ApiError>

State of the unrewarded relayers set at given lane.