Trait t3rn_primitives::bridges::chain_circuit::ToCircuitOutboundLaneApi

pub trait ToCircuitOutboundLaneApi<Block: BlockT, OutboundMessageFee: Parameter, OutboundPayload: Parameter>: Core<Block> {
    fn estimate_message_delivery_and_dispatch_fee(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        lane_id: LaneId,
        payload: OutboundPayload
    ) -> Result<Option<OutboundMessageFee>, ApiError> { ... }
    fn estimate_message_delivery_and_dispatch_fee_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        lane_id: LaneId,
        payload: OutboundPayload
    ) -> Result<Option<OutboundMessageFee>, ApiError> { ... }
    fn messages_dispatch_weight(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        lane: LaneId,
        begin: MessageNonce,
        end: MessageNonce
    ) -> Result<Vec<(MessageNonce, Weight, u32)>, ApiError> { ... }
    fn messages_dispatch_weight_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        lane: LaneId,
        begin: MessageNonce,
        end: MessageNonce
    ) -> Result<Vec<(MessageNonce, Weight, u32)>, ApiError> { ... }
    fn latest_received_nonce(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        lane: LaneId
    ) -> Result<MessageNonce, ApiError> { ... }
    fn latest_received_nonce_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        lane: LaneId
    ) -> Result<MessageNonce, ApiError> { ... }
    fn latest_generated_nonce(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        lane: LaneId
    ) -> Result<MessageNonce, ApiError> { ... }
    fn latest_generated_nonce_with_context(
        &self,
        __runtime_api_at_param__: &BlockId<Block>,
        context: ExecutionContext,
        lane: LaneId
    ) -> Result<MessageNonce, ApiError> { ... }
}

Outbound message lane API for messages that are sent to Circuit chain.

This API is implemented by runtimes that are sending messages to Circuit chain, not the Circuit runtime itself.

Provided Methods

fn estimate_message_delivery_and_dispatch_fee(
&self,
__runtime_api_at_param__: &BlockId<Block>,
lane_id: LaneId,
payload: OutboundPayload
) -> Result<Option<OutboundMessageFee>, ApiError>

Estimate message delivery and dispatch fee that needs to be paid by the sender on this chain.

Returns None if message is too expensive to be sent to Circuit from this chain.

Please keep in mind that this method returns lowest message fee required for message to be accepted to the lane. It may be good idea to pay a bit over this price to account future exchange rate changes and guarantee that relayer would deliver your message to the target chain.

fn estimate_message_delivery_and_dispatch_fee_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
lane_id: LaneId,
payload: OutboundPayload
) -> Result<Option<OutboundMessageFee>, ApiError>

Estimate message delivery and dispatch fee that needs to be paid by the sender on this chain.

Returns None if message is too expensive to be sent to Circuit from this chain.

Please keep in mind that this method returns lowest message fee required for message to be accepted to the lane. It may be good idea to pay a bit over this price to account future exchange rate changes and guarantee that relayer would deliver your message to the target chain.

fn messages_dispatch_weight(
&self,
__runtime_api_at_param__: &BlockId<Block>,
lane: LaneId,
begin: MessageNonce,
end: MessageNonce
) -> Result<Vec<(MessageNonce, Weight, u32)>, ApiError>

Returns total dispatch weight and encoded payload size of all messages in given inclusive range.

If some (or all) messages are missing from the storage, they’ll also will be missing from the resulting vector. The vector is ordered by the nonce.

fn messages_dispatch_weight_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
lane: LaneId,
begin: MessageNonce,
end: MessageNonce
) -> Result<Vec<(MessageNonce, Weight, u32)>, ApiError>

Returns total dispatch weight and encoded payload size of all messages in given inclusive range.

If some (or all) messages are missing from the storage, they’ll also will be missing from the resulting vector. The vector is ordered by the nonce.

fn latest_received_nonce(
&self,
__runtime_api_at_param__: &BlockId<Block>,
lane: LaneId
) -> Result<MessageNonce, ApiError>

Returns nonce of the latest message, received by bridged chain.

fn latest_received_nonce_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
lane: LaneId
) -> Result<MessageNonce, ApiError>

Returns nonce of the latest message, received by bridged chain.

fn latest_generated_nonce(
&self,
__runtime_api_at_param__: &BlockId<Block>,
lane: LaneId
) -> Result<MessageNonce, ApiError>

Returns nonce of the latest message, generated by given lane.

fn latest_generated_nonce_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
lane: LaneId
) -> Result<MessageNonce, ApiError>

Returns nonce of the latest message, generated by given lane.