Module t3rn_primitives::bridges::test_utils

Utilities for testing runtime code.

Structs

Account

A test account which can be used to sign messages.

JustificationGeneratorParams

Configuration parameters when generating test GRANDPA justifications.

Constants

ALICE

Set of test accounts with friendly names.

BOB

CHARLIE

DAVE

EVE

FERDIE

TEST_GRANDPA_ROUND

TEST_GRANDPA_SET_ID

Functions

accounts

Get a list of “unique” accounts.

authorities

Convenience function to get a list of Grandpa authority addresses.

authority_list

Convenience function to get a list of Grandpa authorities.

header_id

Convenience function for generating a Header ID at a given block number.

make_default_justification

Make a valid GRANDPA justification with sensible defaults

make_justification_for_header

Generate justifications in a way where we are able to tune the number of pre-commits and vote ancestries which are included in the justification.

signed_precommit

Create signed precommit with given target.

test_header

Get a header for testing.

test_header_with_correct_parent

Get a header for testing, with correct parent_hash

test_keyring

Get the corresponding identities from the keyring for the “standard” authority set.

voter_set

Get a valid set of voters for a Grandpa round.